fixed problem with non-integral padding in fftPad
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@3024 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Base/scimath/trunk/utils/PaddingUtils.h
@@ -127,7 +127,7 @@ struct PaddingUtils {
   /// @param[in] in input array 
   /// @param[in] out output array (should already be resized to a desired size) 
   /// @param[in] factor additional padding of the output array
-  static void fftPad(const casa::Array<double>& in, casa::Array<double>& out, int factor);
+  static void fftPad(const casa::Array<double>& in, casa::Array<double>& out, float factor);
   
 };
